Shows title of deck when searching for flashcard
Did lots of things. When searching for a flashcard, the title of the deck that houses that flashcard will be shown. Fixed a bug that redirected the user to the wrong location after editing a flashcard. Now when viewing the detail view of a given deck, it is prettier and if it is the author viewing it, then they have the edit/add flashcards/... buttons.