Fixed bug with autonote
If the cursor was at the second paragraph or greater, it crashed when resetting everything because that paragraph no longer existed.  So I wanted to just call an easy function to move the cursor to the beginning, but that doesn't exist so I spent over an hour looking through the horrible SlateJS documentation to find a single function for this.  I finally found some random page that tells you how to move the cursor.